modified:   objects/Disc.cpp 	modified:   objects/Disc.cpp~ 	modified:   objects/Disc.h 	modified:   objects/Disc.h~ 	modified:   tutorials/lesson10/lesson10 	modified:   tutorials/lesson10/lesson10.cpp 	modified:   tutorials/lesson7/lesson07 	modified:   tutorials/lesson7/lesson07.cpp
Added in the physics to Disc's letFly() method. Takes in a power and angle of attack, calculates initial vx and vy, uses default x and y values. Prints out where the disc will land and what it's speed will be when it hits. Driver program (main.cpp) shows an example.